feat: 完成

+++ pkg/operation/types.go
@@ -2,13 +2,13 @@ package operation
 
 // 流程处理参数
 type processArgs struct {
-	StartTime     *string `json:"startTime"`
-	Location      *string `json:"location"`
-	CancelReason1 *string `json:"cancelReason1"`
-	CancelReason2 *string `json:"cancelReason2"`
-	RatingScore   *int    `json:"ratingScore"`
-	RatingText    *string `json:"ratingText"`
-	Letter        *string `json:"letter"`
+	StartTime     *string  `json:"startTime"`
+	Location      *string  `json:"location"`
+	CancelReason1 *string  `json:"cancelReason1"`
+	CancelReason2 *string  `json:"cancelReason2"`
+	RatingScore   *float64 `json:"ratingScore"`
+	RatingText    *string  `json:"ratingText"`
+	Letter        *string  `json:"letter"`
 }
 
 // 新增问答Form
